Corrections to landscape generator bug and removal of warnings
Landscape generator was wrongly setting habitat K for a random continuous landscape.
Also, various changes have been made (especially castings from double to float) to reduce the number of warning messages produced by VS and Linus compilers.

+++ src/Common/Individual.cpp
@@ -935,9 +935,9 @@ else {
 //	<< endl;
 #endif
 emigtraits = new emigTraits;
-emigtraits->d0 = e.d0*eparams.d0Scale + eparams.d0Mean;
-emigtraits->alpha = e.alpha*eparams.alphaScale + eparams.alphaMean;
-emigtraits->beta = e.beta*eparams.betaScale + eparams.betaMean;
+emigtraits->d0 = (float)(e.d0*eparams.d0Scale + eparams.d0Mean);
+emigtraits->alpha = (float)(e.alpha*eparams.alphaScale + eparams.alphaMean);
+emigtraits->beta = (float)(e.beta*eparams.betaScale + eparams.betaMean);
 #if RSDEBUG
 //DEBUGLOG << "Individual::setEmigTraits(): indId=" << indId
 //	<< " emigtraits->d0=" << emigtraits->d0
@@ -1040,9 +1040,9 @@ else {
 	kparams = pSpecies->getKernParams(0,0);
 }
 kerntraits = new trfrKernTraits;
-kerntraits->meanDist1 = k.meanDist1*kparams.dist1Scale + kparams.dist1Mean;
-kerntraits->meanDist2 = k.meanDist2*kparams.dist2Scale + kparams.dist2Mean;
-kerntraits->probKern1 = k.probKern1*kparams.PKern1Scale + kparams.PKern1Mean;
+kerntraits->meanDist1 = (float)(k.meanDist1*kparams.dist1Scale + kparams.dist1Mean);
+kerntraits->meanDist2 = (float)(k.meanDist2*kparams.dist2Scale + kparams.dist2Mean);
+kerntraits->probKern1 = (float)(k.probKern1*kparams.PKern1Scale + kparams.PKern1Mean);
 #if RSDEBUG
 //DEBUGLOG << "Individual::setKernTraits(): indId=" << indId
 //	<< " kerntraits->meanDist1=" << kerntraits->meanDist1
@@ -1098,42 +1098,42 @@ void Individual::setSMSTraits(Species *pSpecies,short SMSgenelocn,short nSMStrai
 //	<< endl;
 #endif
 trfrSMSTraits s = pSpecies->getSMSTraits();
-float dp,gb,alphaDB,betaDB;
+double dp,gb,alphaDB,betaDB;
 dp = gb = alphaDB = betaDB = 0.0;
 if (pGenome != 0) {
 	if (pSpecies->has1ChromPerTrait()) {
 		if (sexdep) {
-			dp = (float)pGenome->express(SMSgenelocn,0,0);
-			gb = (float)pGenome->express(SMSgenelocn+1,0,0);
+			dp = pGenome->express(SMSgenelocn,0,0);
+			gb = pGenome->express(SMSgenelocn+1,0,0);
 			if (nSMStraits == 4) {
-				alphaDB = (float)pGenome->express(SMSgenelocn+2,0,0);
-				betaDB  = (float)pGenome->express(SMSgenelocn+3,0,0);
+				alphaDB = pGenome->express(SMSgenelocn+2,0,0);
+				betaDB  = pGenome->express(SMSgenelocn+3,0,0);
 			}
 		}
 		else {
-			dp = (float)pGenome->express(SMSgenelocn,0,0);
-			gb = (float)pGenome->express(SMSgenelocn+1,0,0);
+			dp = pGenome->express(SMSgenelocn,0,0);
+			gb = pGenome->express(SMSgenelocn+1,0,0);
 			if (nSMStraits == 4) {
-				alphaDB = (float)pGenome->express(SMSgenelocn+2,0,0);
-				betaDB  = (float)pGenome->express(SMSgenelocn+3,0,0);
+				alphaDB = pGenome->express(SMSgenelocn+2,0,0);
+				betaDB  = pGenome->express(SMSgenelocn+3,0,0);
 			}
 		}
 	}
 	else {
 		if (sexdep) {
-			dp = (float)pGenome->express(pSpecies,SMSgenelocn);
-			gb = (float)pGenome->express(pSpecies,SMSgenelocn+1);
+			dp = pGenome->express(pSpecies,SMSgenelocn);
+			gb = pGenome->express(pSpecies,SMSgenelocn+1);
 			if (nSMStraits == 4) {
-				alphaDB = (float)pGenome->express(pSpecies,SMSgenelocn+2);
-				betaDB  = (float)pGenome->express(pSpecies,SMSgenelocn+3);
+				alphaDB = pGenome->express(pSpecies,SMSgenelocn+2);
+				betaDB  = pGenome->express(pSpecies,SMSgenelocn+3);
 			}
 		}
 		else {
-			dp = (float)pGenome->express(pSpecies,SMSgenelocn);
-			gb = (float)pGenome->express(pSpecies,SMSgenelocn+1);
+			dp = pGenome->express(pSpecies,SMSgenelocn);
+			gb = pGenome->express(pSpecies,SMSgenelocn+1);
 			if (nSMStraits == 4) {
-				alphaDB = (float)pGenome->express(pSpecies,SMSgenelocn+2);
-				betaDB  = (float)pGenome->express(pSpecies,SMSgenelocn+3);
+				alphaDB = pGenome->express(pSpecies,SMSgenelocn+2);
+				betaDB  = pGenome->express(pSpecies,SMSgenelocn+3);
 			}
 		}
 	}
@@ -1152,10 +1152,10 @@ if (sexdep) {
 else {
 	smsparams = pSpecies->getSMSParams(0,0);
 }
-smsData->dp = dp*smsparams.dpScale + smsparams.dpMean;
-smsData->gb = gb*smsparams.gbScale + smsparams.gbMean;
+smsData->dp = (float)(dp*smsparams.dpScale + smsparams.dpMean);        
+smsData->gb = (float)(gb*smsparams.gbScale + smsparams.gbMean);
 if (s.goalType == 2) {
-	smsData->alphaDB = alphaDB*smsparams.alphaDBScale + smsparams.alphaDBMean;
+	smsData->alphaDB = (float)(alphaDB*smsparams.alphaDBScale + smsparams.alphaDBMean);
 	smsData->betaDB  = (int)(betaDB*smsparams.betaDBScale + smsparams.betaDBMean + 0.5);
 }
 else {
@@ -1209,7 +1209,7 @@ void Individual::setCRWTraits(Species *pSpecies,short CRWgenelocn,short nCRWtrai
 //	<< " CRWgenelocn=" << CRWgenelocn << " nCRWtraits=" << nCRWtraits << " sexdep=" << sexdep
 //	<< endl;
 #endif
-trfrCRWTraits c; c.stepLength = c.rho = 0.0;
+trfrCRWTraits c; c.stepLength = c.rho = 0.0;          
 if (pGenome != 0) {
 	if (pSpecies->has1ChromPerTrait()) {
 		if (sexdep) {
@@ -1245,8 +1245,8 @@ if (sexdep) {
 else {
 	cparams = pSpecies->getCRWParams(0,0);
 }
-crw->stepL = c.stepLength*cparams.stepLScale + cparams.stepLgthMean;        
-crw->rho   = c.rho*cparams.rhoScale + cparams.rhoMean;
+crw->stepL = (float)(c.stepLength*cparams.stepLScale + cparams.stepLgthMean);        
+crw->rho   = (float)(c.rho*cparams.rhoScale + cparams.rhoMean);
 #if RSDEBUG
 //DEBUGLOG << "Individual::setCRWTraits(): indId=" << indId
 //	<< " crw->stepL=" << crw->stepL	<< " crw->rho=" << crw->rho
@@ -1293,7 +1293,7 @@ void Individual::setSettTraits(Species *pSpecies,short settgenelocn,short nsettt
 //	<< endl;
 #endif
 //simParams sim = paramsSim->getSim();
-settleTraits s; s.s0 = s.alpha = s.beta = 0.0;
+settleTraits s; s.s0 = s.alpha = s.beta = 0.0;            
 if (pGenome != 0) {
 	if (pSpecies->has1ChromPerTrait()) {
 		if (sexdep) {
@@ -1341,9 +1341,9 @@ else {
 //	<< endl;
 #endif
 setttraits = new settleTraits;
-setttraits->s0    = s.s0*sparams.s0Scale + sparams.s0Mean;
-setttraits->alpha = s.alpha*sparams.alphaSScale + sparams.alphaSMean;
-setttraits->beta  = s.beta*sparams.betaSScale + sparams.betaSMean;
+setttraits->s0    = (float)(s.s0*sparams.s0Scale + sparams.s0Mean);
+setttraits->alpha = (float)(s.alpha*sparams.alphaSScale + sparams.alphaSMean);
+setttraits->beta  = (float)(s.beta*sparams.betaSScale + sparams.betaSMean);
 #if RSDEBUG
 //DEBUGLOG << "Individual::setSettTraits(): indId=" << indId
 //	<< " setttraits->s0=" << setttraits->s0
@@ -1544,8 +1544,8 @@ void Individual::moveto(Cell *newCell) {
 // check that location is indeed a neighbour of the current cell
 locn currloc = pCurrCell->getLocn();
 locn newloc = newCell->getLocn();
-double d = sqrt((((double)currloc.x- (double)newloc.x)*((double)currloc.x- (double)newloc.x)
-	+ ((double)currloc.y- (double)newloc.y)*((double)currloc.y- (double)newloc.y)));
+double d = sqrt(((double)currloc.x-(double)newloc.x)*((double)currloc.x-(double)newloc.x)
+	+ ((double)currloc.y-(double)newloc.y)*((double)currloc.y-(double)newloc.y));
 if (d >= 1.0 && d < 1.5) { // ok
 	pCurrCell = newCell; status = 5;
 }
@@ -1967,8 +1967,7 @@ if (status < 7) {
 	double dispmort;
 	trfrMortParams mort = pSpecies->getMortParams();
 	if (trfr.distMort) {
-//		dispmort = 1.0 / (1.0 + std::exp(-(dist - mort.mortBeta)*mort.mortAlpha));
-		dispmort = 1.0 / (1.0 + exp(-(dist - mort.mortBeta)*mort.mortAlpha));  // for LINUX_CLUSTER
+		dispmort = 1.0 / (1.0 + exp(-(dist - mort.mortBeta)*mort.mortAlpha));  
 	}
 	else {
 		dispmort = mort.fixedMort;
@@ -2743,7 +2742,7 @@ if (loopsteps >= 1000 || pNewCell == 0) {
 }
 else {
 	newcellcost = pNewCell->getCost();
-	move.cost = move.dist*0.5f*((float)cellcost + (float)newcellcost);
+	move.cost = move.dist*0.5f*((float)cellcost + (float)newcellcost);  
 	// make the selected move
 	if ((short)memory.size() == movt.memSize) {
 		memory.pop(); // remove oldest memory element
@@ -2797,7 +2796,7 @@ else { // set up the matrix dependent on relationship of previous location to cu
 //    if (write_out) out<<"step 5"<<endl;
   }
 //  if (write_out) out<<"step 6"<<endl;
-  theta = atan2(((double)x- (double)prev.x),((double)y- (double)prev.y));
+	theta = atan2(((double)x-(double)prev.x),((double)y-(double)prev.y));
 //  if (write_out) out<<"prev.x,prev.y: "<<prev.x<<","<<prev.y<<" theta: "<<theta<<endl;
 	d = calcWeightings(dp,(float)theta);
 
@@ -2850,7 +2849,7 @@ else {
 #endif  // PARTMIGRN 
 	}
 	else // goaltype == 2
-		theta = atan2(((double)x - (double)smsData->goal.x),((double)y - (double)smsData->goal.y));
+		theta = atan2(((double)x -(double)smsData->goal.x),((double)y-(double)smsData->goal.y));
 //  if (write_out) out<<"goalx,goaly: "<<goalx<<","<<goaly<<" theta: "<<theta<<endl;
 	d = calcWeightings(gb,(float)theta);
 }
@@ -2859,26 +2858,26 @@ return d;
 }
 
 // Calculate weightings for neighbouring cells
-array3x3d Individual::calcWeightings(const float base,const float theta) {
+array3x3d Individual::calcWeightings(const double base,const double theta) {
 
 array3x3d d; // 3x3 array indexed from SW corner by xx and yy
 int dx,dy,xx,yy;
 
 double i0 = 1.0; 					// direction of theta - lowest cost bias
-double i1 = (double)base;
-double i2 = (double)base * (double)base;
-double i3 = i2 * (double)base;
-double i4 = i3 * (double)base;		// opposite to theta - highest cost bias
+double i1 = base;
+double i2 = base * base;
+double i3 = i2 * base;
+double i4 = i3 * base;		// opposite to theta - highest cost bias
 
 if (fabs(theta) > 7.0 * PI / 8.0) { dx = 0; dy = -1; }
 else {
 	if (fabs(theta) > 5.0 * PI / 8.0) { dy = -1; if (theta > 0) dx = 1; else dx = -1; }
-  else {
-	if (fabs(theta) > 3.0 * PI / 8.0) { dy = 0; if (theta > 0) dx = 1; else dx = -1; }
-    else {
-	  if (fabs(theta) > PI / 8.0) { dy = 1; if (theta > 0) dx = 1; else dx = -1; }
-      else { dy = 1; dx = 0; }
-    }
+	else {
+		if (fabs(theta) > 3.0 * PI / 8.0) { dy = 0; if (theta > 0) dx = 1; else dx = -1; }
+		else {
+			if (fabs(theta) > PI / 8.0) { dy = 1; if (theta > 0) dx = 1; else dx = -1; }
+			else { dy = 1; dx = 0; }
+		}
   }
 }
 //  if (write_out) out<<"goalx,goaly: "<<goalx<<","<<goaly<<" dx,dy: "<<dx<<","<<dy
@@ -2900,7 +2899,7 @@ if (dx == 0 || dy ==0) { // theta points to a cardinal direction
   }
 }
 else { // theta points to an ordinal direction
-  d.cell[dx+1][-dy+1] = (float)i2; d.cell[-dx+1][dy+1] = (float)i2;
+	d.cell[dx+1][-dy+1] = (float)i2; d.cell[-dx+1][dy+1] = (float)i2;          
   xx = dx+1; if (xx > 1) xx -= 2; d.cell[xx+1][dy+1] = (float)i1;
   yy = dy+1; if (yy > 1) yy -= 2; d.cell[dx+1][yy+1] = (float)i1;
   d.cell[-xx+1][-dy+1] = (float)i3; d.cell[-dx+1][-yy+1] = (float)i3;
@@ -2917,7 +2916,7 @@ array3x3f Individual::getHabMatrix(Landscape *pLand,Species *pSpecies,
 
 array3x3f w; // array of effective costs to be returned
 int ncells,x4,y4;
-float weight,sumweights;
+double weight,sumweights;
 // NW and SE corners of effective cost array relative to the current cell (x,y):
 int xmin = 0,ymin = 0,xmax = 0,ymax = 0;
 int cost,nodatacost,h;
@@ -3022,15 +3021,15 @@ for (int x2=-1; x2<2; x2++) {   // index of relative move in x direction
 					}
 					if (prmethod==2) { // harmonic mean
             if (cost > 0) {
-              w.cell[x2+1][y2+1] += (1/(float)cost);
+							w.cell[x2+1][y2+1] += (1.0f/(float)cost);
               ncells++;
             }
           }
           if (prmethod==3) { // arithmetic mean weighted by inverse distance
             if (cost>0) {
               // NB distance is still given by (x3,y3)
-              weight = 1.0f /(float)sqrt((pow((double)x3,2)+pow((double)y3,2)));
-              w.cell[x2+1][y2+1] += weight*(float)cost;
+							weight = 1.0f /(double)sqrt((pow((double)x3,2)+pow((double)y3,2)));
+              w.cell[x2+1][y2+1] += (float)(weight*(double)cost);
               ncells++; sumweights += weight;
             }
           }
@@ -3048,7 +3047,7 @@ for (int x2=-1; x2<2; x2++) {   // index of relative move in x direction
 			  if (prmethod == 1) w.cell[x2+1][y2+1] /= ncells; // arithmetic mean
 			  if (prmethod == 2) w.cell[x2+1][y2+1] = ncells/w.cell[x2+1][y2+1]; // hyperbolic mean
 			  if (prmethod == 3 && sumweights > 0)
-          w.cell[x2+1][y2+1] /= sumweights; // weighted arithmetic mean
+          w.cell[x2+1][y2+1] /= (float)sumweights; // weighted arithmetic mean
       }
 //#if GO2TARGET
 //      if (targetseen > 0) // target is within PR - set to a very low score
